requirements-locked-full.txt
# 依賴鎖定文件
# 生成時間: 2025-09-24T22:48:22.281800
# 基於 GOOGLE 診斷建議，防止版本衝突
# 總共 284 個包，8 個關鍵包

# ==================== 關鍵包（經過兼容性驗證）====================
python-jose==3.3.0 # ✅ 兼容性驗證通過
redis==5.0.1 # ✅ 兼容性驗證通過
psycopg2-binary==2.9.9 # ✅ 兼容性驗證通過
fastapi==0.104.1 # ✅ 兼容性驗證通過
uvicorn==0.24.0 # ✅ 兼容性驗證通過
pydantic==2.5.0 # ✅ 兼容性驗證通過
sqlalchemy==2.0.23 # ✅ 兼容性驗證通過
aiohttp==3.9.1 # ✅ 兼容性驗證通過

# ==================== 其他依賴包 ====================
absl-py==2.3.1
accelerate==1.10.0
aiofiles==23.2.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.18.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asyncio-throttle==1.0.2
asyncpg==0.29.0
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
bandit==1.8.6
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
bitsandbytes==0.47.0
black==23.11.0
blinker==1.9.0
boolean.py==5.0
bottle==0.13.4
bs4==0.0.2
cachecontrol==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
configobj==5.0.9
contourpy==1.3.3
coverage==7.10.2
cryptography==45.0.6
cssselect==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
cyclonedx-python-lib==9.1.0
databricks-sdk==0.64.0
datasets==2.19.1
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.62.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
ecdsa==0.19.1
email-validator==2.1.0
entrypoints==0.4
fake-useragent==2.2.0
filelock==3.18.0
flake8==6.1.0
flask==3.1.2
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2024.3.1
funcy==2.0
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
gto==1.8.0
gunicorn==21.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
license-expression==30.4.4
lxml==6.0.1
lxml_html_clean==0.4.2
mako==1.3.10
markdown==3.8.2
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.5
mccabe==0.7.0
mdurl==0.1.2
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
multitasking==0.0.12
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
nvidia-ml-py==12.575.51
nvidia-ml-py3==7.352.0
omegaconf==2.3.0
openai==1.12.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
optuna==4.4.0
optuna-dashboard==0.19.0
optuna-integration==4.5.0
orjson==3.11.3
outcome==1.3.0.post0
packageurl-python==0.17.5
packaging==25.0
pandas==2.0.3
parse==1.20.2
passlib==1.7.4
pathspec==0.12.1
pbr==7.0.0
peewee==3.18.2
peft==0.17.0
pillow==10.0.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus-client==0.19.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
py-serializable==2.1.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.23.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydot==4.0.1
pyee==13.0.0
pyflakes==3.1.0
pygit2==1.18.2
pygments==2.19.2
pygtrie==2.5.0
pyjwt==2.8.0
pynvml==12.0.0
pyparsing==3.2.3
pyppeteer==0.0.25
pyquery==2.0.1
pysocks==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-oauth2==1.1.1
pytz==2025.2
pywin32==311
pyyaml==6.0.2
regex==2025.7.34
requests==2.31.0
requests-html==0.10.0
rich==13.7.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.13.1
scmrepo==3.5.2
seaborn==0.13.2
selenium==4.35.0
semver==3.0.4
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
sqltrie==0.11.2
starlette==0.27.0
stevedore==5.4.1
structlog==23.2.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.55.0
trio==0.30.0
trio-websocket==0.12.2
trl==0.21.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
w3lib==2.3.1
waitress==3.0.2
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.28
zc.lockfile==3.0.post1
zipp==3.23.0